Add certificate deployment hook, ansible defined.
letsencrypt.sh use umask 077 so usage of group www-data is useless.

The Certificates must be readable for services like apache or dovecot.
But only the specific services should be allowed to read the certificate
for this service. So we must change the owner/group to a specific value
for each certificate. For security reasons this can be only done by root
user.

To have a generic solution a new ansible variable `lcsh_deploy_cert`
exists. This variable must contain bash script for certificate
deployments. Typical tasks on deployment are copy certificate to other
directories, change file owner/permissions and restart services.